% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{magicTokenizer}
\alias{magicTokenizer}
\title{magicTokenizer}
\usage{
magicTokenizer(text)
}
\arguments{
\item{text}{A character string: the text for which the number of tokens is required.}
}
\value{
Returns the number of tokens in the provided text.
}
\description{
This function provides the number of tokens in a specified text, acting as a wrapper for an internal tokenizer function.
}
