% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExperiment.R
\name{runExperiment}
\alias{runExperiment}
\title{Run an Experiment Based on the Configuration}
\usage{
runExperiment(gptConfig, savePath = "./output.xlsx")
}
\arguments{
\item{gptConfig}{A list containing the configuration for the language model, including the system prompt,
model specifications, and token settings.}

\item{savePath}{The file path where the experiment results will be saved in Excel format.
Defaults to './output.xlsx' in the current working directory.}
}
\value{
This function does not return a value but saves the experiment results to the specified Excel file.
Upon completion, "Done." will be printed to the console.
}
\description{
Executes the experiment and saves the results to an Excel file.
}
\examples{
\dontrun{

runExperiment(Experiment_config,"./output.xlsx")

#The first argument Experiment_config is generated by preCheck() function.

Experiment_config <- preCheck(data)
}
}
