% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{setKey}
\alias{setKey}
\title{Step1: Set model's API key and url.}
\usage{
setKey(api_key, api_url = "https://api.openai.com/v1/chat/completions")
}
\arguments{
\item{api_key}{A character string: the user's OpenAI/huggingface/other API key.}

\item{api_url}{A character string: the user's OpenAI/huggingface/other url .default is OpenAI.}
}
\value{
Prints a message to the console indicating whether the API key setup was successful.
If the setup fails, the function stops with an error message.
}
\description{
This function allows users to set and verify an API key for data collection. You can change the default api_url for open-source models' API.
}
\examples{
\dontrun{
set_key(api_key="YOUR_API_KEY", api_url="https://api.openai.com/v1/chat/completions")


}
}
