% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Step2: Load and format data}
\usage{
loadData(
  runList,
  itemIDList,
  conditionList,
  beforePrompt = "",
  targetPrompt,
  afterPrompt = "",
  header = T
)
}
\arguments{
\item{runList}{A numeric vector of data representing the 'Run' column in the experiment.}

\item{itemIDList}{A numeric vector of data representing the 'itemID' column in the experiment.}

\item{conditionList}{A numeric/character vector of data representing the 'Condition' column in the experiment.}

\item{beforePrompt}{A string that will be used as the prefix of the 'Prompt' text.}

\item{targetPrompt}{A character vector of the main prompt (usually experiment items) that will be placed between the beforePrompt and afterPrompt.}

\item{afterPrompt}{An optional string that will be used as the suffix of the 'Prompt' text (default is an empty string).}

\item{header}{A logical value indicating if the output data.frame should include column headers (default is TRUE).}
}
\value{
A data frame with the processed columns 'Run', 'Trial', 'Item', 'Condition', and 'Prompt', ready for use in experiments.
}
\description{
Prepares the stimuli data for the experiment.
}
\examples{
{

df <- data.frame(
Run = c(1, 2, 3, 4, 5, 6, 7, 8, 9),
Item = c(1,1, 1, 1, 1, 1, 1, 1, 2),
Condition = c(1, 2, 3, 4, 5, 6, 7, 8, 2),
TargetPrompt1 = c("The racing driver showed the torn overall...",
                "The racing driver showed the helpful mechanic...",
                "The report claimed that the racing driver showed the torn overall...",
                "The report claimed that the racing driver showed the helpful mechanic...",
                "The racing driver showed the torn overall...",
                "The racing driver showed the helpful mechanic...",
                "The report claimed that the racing driver showed the torn overall...",
                "The report claimed that the racing driver showed the helpful mechanic...",
                "The cricket player showed the ball..."),
TargetPrompt2 = c("The patient showed...",
                 "The patient showed...",
                 "The patient showed...",
                 "The patient showed...",
                 "The rumours alleged that the patient showed...",
                 "The rumours alleged that the patient showed...",
                 "The rumours alleged that the patient showed...",
                 "The rumours alleged that the patient showed...",
                 "The car mechanic showed...")
)
ExperimentItem=loadData(runList=df$Run,itemIDList=df$Item,conditionList=df$Condition,beforePrompt="",targetPrompt = paste(df$TargetPrompt1,df$TargetPrompt2,sep="\n"),afterPrompt="")
print(ExperimentItem)
}

}
